Add and reorganize e2e test utils
