+++ tests/e2e/utils/index.js
@@ -1,0 +1,4 @@
+export { activatePlugin } from './activate-plugin';
+export { deactivatePlugin } from './deactivate-plugin';
+export { activateExperience } from './activate-experience';
+export { deactivateExperience } from './deactivate-experience';
